Battle ready update (#76)
* first playable state, WIP

* fixed extra setup

* small changes

* moved the setup states to better spots

* refactor

* fixed prompt distribution

* fixed some timers

* fixed game speed naming

* temporarily removed extra setup drawing

* fixed extra setup logic bug

* Body builder commented changes

* Update ExtraSetupDrawing_GS.cs

* Update EnumerableExtensions.cs

* Update ExtraSetupDrawing_GS.cs

+++ RoystonGame/TV/GameModes/KevinsGames/Mimic/DataModels/MimicConstants.cs
@@ -16,7 +16,7 @@ namespace RoystonGame.TV.GameModes.KevinsGames.Mimic.DataModels
         {
             return numberOfPLayersInLobby * 20;
         }
-        public const double MimicTimerMultiplier = 2.0;
+        public const float MimicTimerMultiplier = 2.0f;
 
         public const double MemorizeTimerLength = 10;
